Add Skill entity, related DTOs, form, service and repository
This commit introduces the Skill entity with associated Grade enum. Also added are the DTOs SkillFullDto and SkillShortDto, a form class SkillForm, the service interface SkillService and its implementation SkillServiceImpl, and the repository SkillRepositories. These provide a blueprint for managing skill-related data within the application. The database connection string in application.yml has also been changed to a "pokesleep" database.